requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
azure-core==1.1.1
azure-storage-blob==12.1.0
azure-storage-file-datalake==12.0.0b6
backcall==0.1.0
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
fsspec==0.6.2
idna==2.8
importlib-metadata==1.3.0
ipython==7.10.2
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.15.2
more-itertools==8.0.2
msrest==0.6.10
oauthlib==3.1.0
packaging==19.2
parso==0.5.2
pathspec==0.6.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
pytest==5.3.2
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
